Update F1toPCO/Models/People.cs
The issue is related to the idField (file: People.cs, ~line 729 
& ~line 791). The code currently uses a short to hold the
idField... well the issue with that is that a short can
only hold values from -32768 to 32767 (16 bits).

It turns out that one of the F1 records that we were updating
had an idField value of greater than 32767, and it was causing
the sync tool to crash when it was trying to sync the record. 
I was able to determine this until I put a try/catch block in
the function that gets F1 records by the last updated date.

So in our fork I've changed the idField from a short to an int... 
so now we can have ids from -2,147,483,648 to 2,147,483,647.